src/functions/image/requirements.txt
aws-lambda-typing==2.20.0
boto3==1.42.5
pillow==12.0.0
